exclude firewalled and local from travis
