+++ pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 markers = 
-  localtest: mark for local testing only, skip Travis
+	local: locally run tests to be excluded from Travis
+  	firewalled: tests that can only run inside firewall network, skip Travis
